+++ components/form-builder/__tests__/useActivePathname.tsx
@@ -5,9 +5,6 @@ it("removes id from path", () => {
   expect(cleanPath(`/form-builder/settings/clawwjklm05188cygx45fkihp`)).toBe(
     "/form-builder/settings"
   );
-  expect(cleanPath(`/form-builder/preview/settings/clatrihnk19288ayfb6zpg53j`)).toBe(
-    "/form-builder/preview/settings"
-  );
   expect(cleanPath(`/form-builder/edit/clatrihnk19288ayfb6zpg53j`)).toBe("/form-builder/edit");
   expect(cleanPath()).toBe("");
   expect(new URL("https://example.com/settings", "https://example.com").pathname).toBe("/settings");
