+++ components/form-builder/app/TestDataDelivery.tsx
@@ -67,7 +67,7 @@ export const TestDataDelivery = () => {
           className="text-underline inline"
           onClick={(e) => {
             e.preventDefault();
-            router.push({ pathname: `/form-builder/preview/settings` });
+            router.push({ pathname: `/form-builder/settings` });
           }}
         >
           {t("updateResponseDestination")}
