+++ components/form-builder/app/shared/CopyToClipboard.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useTranslation } from "next-i18next";
 
-import { useTemplateStore } from "../store/useTemplateStore";
-import { Button, withMessage } from "./shared/Button";
+import { useTemplateStore } from "../../store/useTemplateStore";
+import { Button, withMessage } from "./Button";
 
 export const CopyToClipboard = () => {
   const { t } = useTranslation("form-builder");
