+++ components/form-builder/app/shared/DownloadFileButton.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback } from "react";
 import { useTranslation } from "next-i18next";
 
-import { useTemplateStore } from "../store/useTemplateStore";
-import { Button } from "./shared/Button";
+import { useTemplateStore } from "../../store/useTemplateStore";
+import { Button } from "./Button";
 
 const slugify = (str: string) =>
   str
