+++ components/form-builder/hooks/index.ts
@@ -2,3 +2,4 @@ export { useAllowPublish } from "./useAllowPublish";
 export { useDeleteForm } from "./useDelete";
 export { useElementOptions } from "./useElementOptions";
 export { usePublish } from "./usePublish";
+export { useActivePathname } from "./useActivePathname";
